BayesianSmartContract.sol
BayesianSmartContract is an on-chain implementation of Bayesian inference in Solidity that classifies physician visits as in-person (PPH) or remote (PPR) based on multiple real-world evidence inputs. Unlike traditional smart contracts that rely on external oracles to provide and interpret data, this contract embeds a probabilistic reasoning engine directly within the Ethereum Virtual Machine (EVM), enabling trustless, transparent, and automated decision-making based on real-world medical indicators. It is trustworthy due to the provable data with evidence.